Update ingredient amount type

+++ exzellenzkoch/src/main/kotlin/de/psekochbuch/exzellenzkoch/application/dto/IngredientDto.kt
@@ -17,7 +17,7 @@ data class IngredientDto(
         /**
          * Amount of the ingredient
          */
-        var amount:Int,
+        var amount:Double,
 
         /**
          * Unit of the ingredient
